fix bag db

+++ api/src/main/java/ru/gb/springShop/api/CartItemDto.java
@@ -1,6 +1,7 @@
 package ru.gb.springShop.api;
 
 
+import java.math.BigDecimal;
 
 //элемент корзины
 public class CartItemDto {
@@ -8,8 +9,8 @@ public class CartItemDto {
     private Long productId;
     private String productTitle;
     private int quantity;//количество
-    private int pricePerProduct;
-    private int price;
+    private BigDecimal  pricePerProduct;
+    private BigDecimal  price;
 
     public Long getProductId() {
         return productId;
@@ -35,19 +36,19 @@ public class CartItemDto {
         this.quantity = quantity;
     }
 
-    public int getPricePerProduct() {
+    public BigDecimal  getPricePerProduct() {
         return pricePerProduct;
     }
 
-    public void setPricePerProduct(int pricePerProduct) {
+    public void setPricePerProduct(BigDecimal  pricePerProduct) {
         this.pricePerProduct = pricePerProduct;
     }
 
-    public int getPrice() {
+    public BigDecimal getPrice() {
         return price;
     }
 
-    public void setPrice(int price) {
+    public void setPrice(BigDecimal  price) {
         this.price = price;
     }
 }
