+++ api/src/main/java/ru/gb/springShop/api/ProductDto.java
@@ -5,10 +5,12 @@ package ru.gb.springShop.api;
 // data transfer object (структура для передачи)
 //на фронт всегда отдаем ДТО. Ломая в ДБ необязательные поля, на стороне фронта изминений не будет.
 
+import java.math.BigDecimal;
+
 public class ProductDto {
     private Long id;
     private String title;
-    private int price;
+    private BigDecimal price;
 
     public Long getId() {
         return id;
@@ -26,15 +28,15 @@ public class ProductDto {
         this.title = title;
     }
 
-    public int getPrice() {
+    public BigDecimal  getPrice() {
         return price;
     }
 
-    public void setPrice(int price) {
+    public void setPrice(BigDecimal  price) {
         this.price = price;
     }
 
-    public ProductDto(Long id, String title, int price) {
+    public ProductDto(Long id, String title, BigDecimal  price) {
         this.id = id;
         this.title = title;
         this.price = price;
